+++ lang/en/filter_filtercodes.php
@@ -42,6 +42,15 @@ $string['enable_scrape_description'] = 'Enable the scrape tag.';
 $string['escapebraces'] = 'Escape tags';
 $string['escapebraces_desc'] = 'When this option is checked, you will be able to display FilterCode tags without them being interpreted by this filter by wrapping your tag in [ brackets ]. This can be very useful when creating FilterCodes documentation for the teachers and course creators on your Moodle site.<br><br>Example: [{fullname}] will not display the user\'s full name but display the {fullname} tag instead without the brackets.';
 
+$string['sizeb'] = 'B';
+$string['sizekb'] = 'KB';
+$string['sizemb'] = 'MB';
+$string['sizegb'] = 'GB';
+$string['sizetb'] = 'TB';
+$string['sizeeb'] = 'EB';
+$string['sizezb'] = 'ZB';
+$string['sizeyb'] = 'YB';
+
 $string['formquickquestion'] = '
 <form action="{wwwroot}/local/contact/index.php" method="post" class="cf contact-us">
     <fieldset>